+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -100,6 +100,8 @@
     <Compile Include="EventWarlock\WarlockAction.cs" />
     <Compile Include="EventWarlock\WarlockActions\WarlockActionEcho.cs" />
     <Compile Include="EventWarlock\WarlockConditions\FirstEqualsSecond.cs" />
+    <Compile Include="Extensions\ByteArrayComparer.cs" />
+    <Compile Include="Extensions\NullableByteArrayComparer.cs" />
     <Compile Include="Memory\FileInterface.cs" />
     <Compile Include="Memory\FileMemoryInterface.cs" />
     <Compile Include="Memory\IMemoryDomain.cs" />
